Milagres da Cintia

+++ src/pages/Timeline/index.js
@@ -2,9 +2,14 @@ import { signOut } from '../../services/index.js';
 
 export default () => {
   const user = firebase.auth().currentUser;
-  const timeline = (document.getElementById("container").innerHTML = `
+ 
+  const timeline = document.createElement('div');
+  // timeline.setAttribute('class', 'container2');
+
+  timeline.innerHTML = `
   <link rel="stylesheet" href="./pages/Timeline/style.css" />
 
+  <div class="banner-menu">
     <div class="banner">
       <img src="assets/logo.png" alt="Logo">
       <div class="title-container">
@@ -21,23 +26,29 @@ export default () => {
     </label>
 
     <ul class="inside-menu">
-      <li>
-        <img id="preview" src="${user.photoURL || '../../assets/default-user-img.png'}" class="user-photo-menu">
-        <input type="file" id="photo" class="inputImg">
-        <button id="uploadImage" class="buttonImg">Alterar Foto Perfil</button>
-      </li>
-      <li>
-        <p class="username-menu"> <b>${user.displayName || "Usuário"} </b> </p>
-      </li>
-      <li>
-        <p class="email-menu"> ${user.email || "Usuário"} </p>
-      </li>
+      <div class="profile-container">
+        <li class="upload-photo">
+
+            <img id="preview" src="${user.photoURL || '../../assets/default-user-img.png'}" class="user-photo-menu">
+            <div class="hidden-div">
+              <input type="file" id="photo" class="inputImg">
+              <button id="uploadImage" class="buttonImg">Alterar Foto Perfil</button>
+            </div>
+          
+        </li>
+        <li>
+          <p class="username-menu"> <b>${user.displayName || "Usuário"} </b> </p>
+        </li>
+        <li>
+          <p class="email-menu"> ${user.email || "Usuário"} </p>
+        </li>
+      </div>
 
       <button id="signout-button" class="signout-button buttons">
         <img src="./assets/exit.png" alt="Ícone de Saída">
       </button>
     </ul>
-
+  </div>
     <form action="" id="postForm">
       <textarea type="textarea" id="postText" class="post-textarea" rows="5" cols="50" placeholder="Digite aqui sua review..."></textarea>
       <button type="submit" class="buttons post-button"> Publicar </button>
@@ -45,10 +56,10 @@ export default () => {
 
     <ul id="posts"></ul>
 
-  `);
+  `;
 
   //Sair da conta do usuário
-  document.getElementById("signout-button").addEventListener("click", (e) => {
+  timeline.querySelector("#signout-button").addEventListener("click", (e) => {
     e.preventDefault();
     signOut();
   });
@@ -57,27 +68,30 @@ export default () => {
   const postsCollection = firebase.firestore().collection("posts");
 
   // Enviando posts para o firestore
-  document.getElementById("postForm").addEventListener("submit", (event) => {
+   timeline.querySelector("#postForm").addEventListener("submit", (event) => {
     event.preventDefault();
+    const text =  timeline.querySelector("#postText").value;
 
-    const text = document.getElementById("postText").value;
-
+    const getDate = () => {
+      const date = new Date();
+      return date.toLocaleString('pt-BR');    
+    };
+  
     const post = {
+      user: firebase.auth().currentUser.email,
       text: text,
-      userId: "Patrícia",
       likes: 0,
-      comments: [],
+      date: getDate(),
     };
 
     postsCollection.add(post).then(() => {
-      document.getElementById("postText").value = "";
+       timeline.querySelector("#postText").value = "";
       loadPosts();
     });
   });
 
   // Adicionando posts
   function createTemplatePost(post) {
-    const date = new Date();
     
     const postTemplate = `
       <li class="posts-box">
@@ -86,34 +100,44 @@ export default () => {
             <img src="${user.photoURL || '../../assets/default-user-img.png'}" class="user-photo">
             <div class="username-date-container">
               <p class="username"> ${user.displayName || "Usuário"} </p>
-              <time class="date">${date.toLocaleString('pt-BR')}</time>
+              <time class="date">${post.data().date}</time>
             </div>
           </div>
+        
+          <div id=${post.id}>
+            <textarea disabled class="post-value"> ${post.data().text} </textarea>
+          
+            <button class='save-edit-button buttons display-none' type='button'>Salvar</button>
+          </div>
 
-          <p class="post-value"> ${post.data().text} </p>
-      
-          <div class="like-comment">
-            <p> ❤️ ${post.data().likes} </p>
-            <p> 💬 </p> 
+          <div id=${post.id} class="like-comment">
+            <button class="likePost-btn timeline-buttons">❤️ ${post.data().likes}</button>  
           </div>
         </div>
           
         <div id=${post.id} class="buttons-container">
+          <button class="editPost-btn timeline-buttons">✏️</button>
           <button class="deletePost-btn timeline-buttons">🗑️</button>
-          <button class="likePost-btn timeline-buttons">❤️</button>
-          <button class="commentPost-btn timeline-buttons">💬</button>
-          <div></div>
         </div>
       </li>
     `;
-    document.getElementById("posts").innerHTML += postTemplate;
 
-    // Deletando posts
-    const deleteButtons = document.querySelectorAll(".deletePost-btn");
+    const postBox =  timeline.querySelector("#posts");
+    postBox.innerHTML += postTemplate;
+
+    // Deletar posts
+    const deleteButtons = timeline.querySelectorAll(".deletePost-btn");
     for (const button of deleteButtons) {
       button.addEventListener("click", function (event) {
-        console.log(event.target.parentNode.id);
-        deletePost(event.target.parentNode.id);
+        const deleteConfirmation = confirm("Tem certeza quer deseja deletar esse post?");
+
+        if(deleteConfirmation) {
+          deletePost(event.target.parentNode.id);
+        }
+        else {
+          return false;
+        }
+        
       });
     }
 
@@ -126,13 +150,18 @@ export default () => {
         });
     }
 
-    // Curtindo posts
-    const likeButtons = document.querySelectorAll(".likePost-btn");
+    // Curtir e descurtir posts
+    const likeButtons = timeline.querySelectorAll(".likePost-btn");
 
     for (const button of likeButtons) {
       button.addEventListener("click", function (event) {
-        console.log(event.target.parentNode.id);
-        addLikes(event.target.parentNode.id);
+        if (post.data().likes.length >= 1) {
+          deleteLikes(event.target.parentNode.id)
+          console.log("oi")
+        } else {
+          addLikes(event.target.parentNode.id);
+          console.log("ola")
+        }     
       });
     }
 
@@ -146,14 +175,59 @@ export default () => {
           loadPosts();
         });
     }
+
+    function deleteLikes(id) {
+      postsCollection
+        .doc(id)
+        .update({
+          likes: firebase.firestore.FieldValue.increment(0),
+        })
+        .then(() => {
+          loadPosts();
+        });
+    }
+
+    // Editar post
+    const editButtons = timeline.querySelectorAll(".editPost-btn");
+    
+    for (const button of editButtons) {
+      button.addEventListener("click", function () {
+        openEditPost(postBox);
+      });
+    }
+
+    function openEditPost (element) {
+        element.querySelector('.post-value').removeAttribute('disabled');
+        element.querySelector('.save-edit-button').classList.remove('display-none');
+    }
+
+    // editPost(event.target.parentNode.id);
+    
+    // function editPost (newText, id) {
+    //   postsCollection
+    //     .doc(id)
+    //     .update({ 
+    //       text: newText 
+    //     });
+    // };
+
+    // Visibilidade dos botões de editar e deletar
+    // const visibilityOfButtons = (document, user) => {
+    //   if (user !== firebase.auth().currentUser.email) {
+    //     document.querySelector('#deletePost-btn').classList.add('visibility-hidden');
+    //     document.querySelector('#editPost-btn').classList.add('visibility-hidden');
+    //   }
+    // };
+  
+    // visibilityOfButtons(document, user);
   }
 
   // Adicionando foto do perfil
-  const uploadImage = document.querySelector('#uploadImage');
+  const uploadImage = timeline.querySelector('#uploadImage');
   uploadImage.addEventListener('click', () => {
     console.log('botão de upa img');
     const ref = firebase.storage().ref();
-    const file = document.querySelector('#photo').files[0];
+    const file = timeline.querySelector('#photo').files[0];
     const name = new Date() + '-' + file.name
     const metadata = {
       contentType: file.type,
@@ -164,7 +238,7 @@ export default () => {
       .then(url => {
         console.log(url);
         console.log('imagem upada');
-        const image = document.querySelector('#preview');
+        const image = timeline.querySelector('#preview');
         image.src = url;
         const userUp = firebase.auth().currentUser;
         userUp.updateProfile({
@@ -176,10 +250,10 @@ export default () => {
 
   // Mostrando os posts na tela
   function loadPosts() {
-    document.getElementById("posts").innerHTML = "Carregando posts...";
+     timeline.querySelector("#posts").innerHTML = `<span class="loading-post">Carregando posts...</span>`;
 
     postsCollection.get().then((snap) => {
-      document.getElementById("posts").innerHTML = "";
+       timeline.querySelector("#posts").innerHTML = "";
       snap.forEach((post) => {
         createTemplatePost(post);
       });
